ai-worker/requirements.txt
redis
pymysql
python-dotenv
minio
faster-whisper
openai
pydantic
python-dotenv